move BufferedSliceReader and BufferedSliceWriter into webknossos
